Refactor start battle overlay into modular steps
